fix(friends): lift search logic to screen and fix setResults undefined error
