+++ src/app/(protected)/friends/search.tsx
@@ -1,5 +1,6 @@
 import { useState } from 'react';
 import { View } from 'react-native';
+import { debounce } from 'lodash';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import { Profile } from '@/types';
 import UserProfileModal from '@/components/userProfileModal';
@@ -12,6 +13,9 @@ type SearchResult = Profile & { status: 'accepted' | 'pending' | 'none' };
 export default function FriendsSearchScreen() {
   const { user } = useUser();
   const [selectedUser, setSelectedUser] = useState<SearchResult | null>(null);
+  const [results, setResults] = useState<SearchResult[]>([]);
+  const [query, setQuery] = useState('');
+  const [loading, setLoading] = useState(false);
 
   const handleSendFriendRequest = async (targetUserId: string) => {
     try {
@@ -47,10 +51,54 @@ export default function FriendsSearchScreen() {
     }
   };
 
+  const handleSearchUsers = async (username: string) => {
+    if (!username || username.length < 2) {
+      setResults([]);
+      return;
+    }
+    setLoading(true);
+    try {
+      const token = await AsyncStorage.getItem('access_token');
+      const res = await fetch(
+        `${process.env.EXPO_PUBLIC_API_URL}/friends/search?username=${username}`,
+        {
+          method: 'GET',
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${token}`,
+          },
+        },
+      );
+
+      if (!res.ok) {
+        const { error } = await res.json();
+        console.error('Error fetching users:', error);
+        return;
+      }
+
+      const data = await res.json();
+      setResults(data ?? []);
+    } catch (error) {
+      console.error('Error', error);
+    }
+    setLoading(false);
+  };
+
+  const debouncedSearch = debounce(handleSearchUsers, 300);
+
+  const handleChange = (text: string) => {
+    setQuery(text);
+    debouncedSearch(text);
+  };
+
   return (
     <View className="flex-1 bg-black px-4 pt-12">
       <FriendsSearchList
+        results={results}
+        loading={loading}
         onUserPress={(u) => setSelectedUser(u)}
+        query={query}
+        onChangeQuery={handleChange}
         actionComponent={(u) => (
           <AddFriendButton
             isRequestSent={u.status === 'pending'}
